qcacmn: FT-SAE roam fails due to null TLV in RSO Command
The commit Ibe64fb346a24962e59ca461c3099270d8766adc5 has missed
some changes in 2.0.3 driver, as the propagation did not
included all the required mainline changes.
And due to the missing change the wmi_roam_11r_offload_tlv_param
TLV in RSO command was not filled and hence it resulted in FT-SAE
roaming failure.
This fix brings in the missing changes that was part of this
mainline commit and fixes the FT-SAE roaming.

Change-Id: I29699ced7d9e9bf2d2bf675c49f3239a9e1e67f7
CRs-Fixed: 2737095