qcacld-3.0: Parse FTIE with MIC length 24 for SHA384 AKMs
The commit I5aa50145fcd3ba91b1c92d4817b7f0e4fc216e3f has missed
some changes in 2.0.3 driver, as the propagation did not
included all the required mainline changes.
And due to the missing change the FT-SuiteB roam fails.

This fix brings in the missing changes that was part of this
mainline commit. For SHA384 based AKMs, add changes to reparse
the association/reassociation response FT element. Introduce new
FTIE structure with MIC defined as array of 24 bytes. With this,
the R0KH-ID and R1KH-ID will be populated correctly in to the
assoc response structure and ultimately RSO command will carry
the right R0KH-ID to firmware.

Change-Id: I483652d083f13bb7ce058fe487bd6f04f26a87c2
CRs-Fixed: 2747599